experiments/babel/app/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.3.1
    # via mesop
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   starlette
    #   watchfiles
blinker==1.9.0
    # via flask
cachetools==6.1.0
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   flask
    #   rich-toolkit
    #   typer
    #   uvicorn
dataclasses-json==0.6.7
    # via babel-web (pyproject.toml)
deepdiff==8.6.1
    # via mesop
deprecated==1.2.18
    # via opentelemetry-api
fastapi==0.118.3
    # via babel-web (pyproject.toml)
fastapi-cli==0.0.13
    # via babel-web (pyproject.toml)
flask==3.1.2
    # via
    #   babel-web (pyproject.toml)
    #   mesop
google-api-core==2.25.2
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-auth==2.40.3
    # via
    #   babel-web (pyproject.toml)
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via google-cloud-logging
google-cloud-logging==3.12.1
    # via babel-web (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-logging
grpcio==1.75.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.75.1
    # via google-api-core
gunicorn==23.0.0
    # via babel-web (pyproject.toml)
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   werkzeug
marshmallow==4.0.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mesop==1.1.0
    # via babel-web (pyproject.toml)
msgpack==1.1.1
    # via mesop
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.3
    # via
    #   babel-web (pyproject.toml)
    #   pandas
opentelemetry-api==1.37.0
    # via google-cloud-logging
ordered-set==4.1.0
    # via deepdiff
packaging==25.0
    # via
    #   gunicorn
    #   marshmallow
pandas==2.3.3
    # via babel-web (pyproject.toml)
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   mesop
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.10
    # via
    #   fastapi
    #   mesop
pydantic-core==2.39.0
    # via pydantic
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   mesop
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via google-api-core
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via fastapi-cli
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.49.1
    # via fastapi
typer==0.19.2
    # via fastapi-cli
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2025.2
    # via pandas
urllib3==2.6.0
    # via requests
uvicorn==0.36.1
    # via
    #   babel-web (pyproject.toml)
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchdog==6.0.0
    # via mesop
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
werkzeug==3.1.4
    # via
    #   babel-web (pyproject.toml)
    #   flask
    #   mesop
wrapt==1.17.3
    # via deprecated
zipp==3.23.0
    # via importlib-metadata
